requirements.txt
# Core ML/DL Framework Dependencies
torch==2.2.2
torchvision==0.17.2
transformers==4.38.2
diffusers==0.30.2
accelerate==0.28.0
bitsandbytes==0.43.0
peft==0.15.1

# Face Processing & Computer Vision Core
mediapipe==0.10.21
dlib-bin==19.24.6
opencv-python==4.9.0.80
scikit-image==0.25.2 --only-binary :all:

# 3D Processing & Visualization
open3d==0.19.0
trimesh==4.0.5
pyvista==0.44.1

# Scientific Computing Core
numpy
scipy==1.11.4
scikit-learn>=1.3.0  # for lighting coefficient calculations
matplotlib>=3.7.0    # for visualizations

# Utility Libraries
modal-client==1.0.0
pyyaml==6.0.1